Change: Cleanup dependency_graph
