+++ troubadix/standalone_plugins/dependency_graph/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+# SPDX-FileCopyrightText: 2025 Greenbone AG
